refactor:#166 경로 변경
