+++ src/app/(auth)/find-password/layout.tsx
@@ -1,5 +1,5 @@
-import GlobalNavigationBar from '@/components/layouts/global-navigation-bar';
-import HeaderWithProfile from '@/components/layouts/header-with-profile';
+import GlobalNavigationBar from '@/components/layouts/global-navigation-bar/global-navigation-bar';
+import HeaderWithProfile from '@/components/layouts/header/header-with-profile';
 import Footer from '@/components/commons/footer';
 
 const Layout = ({
