+++ src/app/(client)/my-page/set-goal/layout.tsx
@@ -1,5 +1,5 @@
 import { Suspense } from 'react';
-import Header from '@/components/layouts/header';
+import Header from '@/components/layouts/header/header';
 
 const Layout = ({
   children
