+++ src/app/(home)/layout.tsx
@@ -1,6 +1,6 @@
 import Footer from '@/components/commons/footer';
-import GlobalNavigationBar from '@/components/layouts/global-navigation-bar';
-import HeaderWithProfile from '@/components/layouts/header-with-profile';
+import GlobalNavigationBar from '@/components/layouts/global-navigation-bar/global-navigation-bar';
+import HeaderWithProfile from '@/components/layouts/header/header-with-profile';
 
 const Layout = ({
   children
